+++ test/karma.conf.js
@@ -39,10 +39,14 @@ module.exports = function(config) {
       // endbower
 
       // application
+      'test/mocks/beforeModuleMocks.js',
+      'app/components/state/state.js', // Load these first to prevent dep clash
+      'app/templates.js',
+      'app/components/omnibox/omnibox.js',
       'app/lizard-nxt.js',
-      'app/*.js',
       'test/mocks/**/*.js',
-      'app/lib/*.js',
+      'app/lib/**/*.js',
+      'app/*.js',
       'app/components/**/{controllers/,directives/,services/,*}.js',
       // end application